+++ src/api/style/PageBreak.ts
@@ -1,0 +1,5 @@
+export enum PageBreak {
+  None,
+  Before,
+  After
+}